+++ ThreeDSecure/src/main/java/com/braintreepayments/api/ThreeDSecureV2TextBoxCustomization.java
@@ -12,20 +12,59 @@ import com.cardinalcommerce.shared.userinterfaces.TextBoxCustomization;
  */
 public class ThreeDSecureV2TextBoxCustomization extends ThreeDSecureV2BaseCustomization implements Parcelable {
 
-    private int borderWidth;
-    private String borderColor;
-    private int cornerRadius;
     private final TextBoxCustomization cardinalTextBoxCustomization = new TextBoxCustomization();
 
     public ThreeDSecureV2TextBoxCustomization() {
-        super.cardinalValue = cardinalTextBoxCustomization;
+    }
+
+    /**
+     * @param textFontName Font type for the UI element.
+     */
+    public void setTextFontName(@Nullable String textFontName) {
+        cardinalTextBoxCustomization.setTextFontName(textFontName);
+    }
+
+    /**
+     * @param textColor Color code in Hex format. For example, the color code can be “#999999”.
+     */
+    public void setTextColor(@Nullable String textColor) {
+        cardinalTextBoxCustomization.setTextColor(textColor);
+    }
+
+    /**
+     * @param textFontSize Font size for the UI element.
+     */
+    public void setTextFontSize(int textFontSize) {
+        cardinalTextBoxCustomization.setTextFontSize(textFontSize);
+    }
+
+    /**
+     * @return Font type for the UI element.
+     */
+    @Nullable
+    public String getTextFontName() {
+        return cardinalTextBoxCustomization.getTextFontName();
+    }
+
+    /**
+     * @return Color code in Hex format.
+     */
+    @Nullable
+    public String getTextColor() {
+        return cardinalTextBoxCustomization.getTextColor();
+    }
+
+    /**
+     * @return Font size for the UI element.
+     */
+    public int getTextFontSize() {
+        return cardinalTextBoxCustomization.getTextFontSize();
     }
 
     /**
      * @param borderWidth Width (integer value) of the text box border.
      */
     public void setBorderWidth(int borderWidth) {
-        this.borderWidth = borderWidth;
         cardinalTextBoxCustomization.setBorderWidth(borderWidth);
     }
 
@@ -33,7 +72,6 @@ public class ThreeDSecureV2TextBoxCustomization extends ThreeDSecureV2BaseCustom
      * @param borderColor Color code in Hex format. For example, the color code can be “#999999”.
      */
     public void setBorderColor(@Nullable String borderColor) {
-        this.borderColor = borderColor;
         cardinalTextBoxCustomization.setBorderColor(borderColor);
     }
 
@@ -41,7 +79,6 @@ public class ThreeDSecureV2TextBoxCustomization extends ThreeDSecureV2BaseCustom
      * @param cornerRadius Radius (integer value) for the text box corners.
      */
     public void setCornerRadius(int cornerRadius) {
-        this.cornerRadius = cornerRadius;
         cardinalTextBoxCustomization.setCornerRadius(cornerRadius);
     }
 
@@ -49,7 +86,7 @@ public class ThreeDSecureV2TextBoxCustomization extends ThreeDSecureV2BaseCustom
      * @return Width (integer value) of the text box border.
      */
     public int getBorderWidth() {
-        return borderWidth;
+        return cardinalTextBoxCustomization.getBorderWidth();
     }
 
     /**
@@ -57,14 +94,14 @@ public class ThreeDSecureV2TextBoxCustomization extends ThreeDSecureV2BaseCustom
      */
     @Nullable
     public String getBorderColor() {
-        return borderColor;
+        return cardinalTextBoxCustomization.getBorderColor();
     }
 
     /**
      * @return Radius (integer value) for the text box corners.
      */
     public int getCornerRadius() {
-        return cornerRadius;
+        return cardinalTextBoxCustomization.getCornerRadius();
     }
 
     TextBoxCustomization getCardinalTextBoxCustomization() {
@@ -78,17 +115,40 @@ public class ThreeDSecureV2TextBoxCustomization extends ThreeDSecureV2BaseCustom
 
     @Override
     public void writeToParcel(Parcel parcel, int i) {
-        super.writeToParcel(parcel, i);
-        parcel.writeInt(borderWidth);
-        parcel.writeString(borderColor);
-        parcel.writeInt(cornerRadius);
+        parcel.writeString(cardinalTextBoxCustomization.getTextFontName());
+        parcel.writeString(cardinalTextBoxCustomization.getTextColor());
+        parcel.writeInt(cardinalTextBoxCustomization.getTextFontSize());
+        parcel.writeInt(cardinalTextBoxCustomization.getBorderWidth());
+        parcel.writeString(cardinalTextBoxCustomization.getBorderColor());
+        parcel.writeInt(cardinalTextBoxCustomization.getCornerRadius());
     }
 
     private ThreeDSecureV2TextBoxCustomization(Parcel in) {
-        super(in);
-        borderWidth = in.readInt();
-        borderColor = in.readString();
-        cornerRadius = in.readInt();
+        String textFontName = in.readString();
+        String textColor = in.readString();
+        int textFontSize = in.readInt();
+        int borderWidth = in.readInt();
+        String borderColor = in.readString();
+        int cornerRadius = in.readInt();
+
+        if (textFontName != null) {
+            cardinalTextBoxCustomization.setTextFontName(textFontName);
+        }
+        if (textColor != null) {
+            cardinalTextBoxCustomization.setTextColor(textColor);
+        }
+        if (textFontSize != 0) {
+            cardinalTextBoxCustomization.setTextFontSize(textFontSize);
+        }
+        if (borderWidth != 0) {
+            cardinalTextBoxCustomization.setBorderWidth(borderWidth);
+        }
+        if (borderColor != null) {
+            cardinalTextBoxCustomization.setBorderColor(borderColor);
+        }
+        if (cornerRadius != 0) {
+            cardinalTextBoxCustomization.setCornerRadius(cornerRadius);
+        }
     }
 
     public static final Creator<ThreeDSecureV2TextBoxCustomization> CREATOR = new Creator<ThreeDSecureV2TextBoxCustomization>() {