+++ CHANGELOG.md
@@ -2,12 +2,18 @@
 
 ## unreleased
 
+* ThreeDSecure  
+  * Support AndroidX and remove Jetifier requirement (fixes #315)
+  * Bump Cardinal version to `2.2.6-1`
+  * Fix null pointer error in V2 UI customization
+  * Deprecate `ThreeDSecureV2BaseCustomization`
+  * Deliver browser switch result asynchronously on main thread
+* SamsungPay
+  * Support AndroidX and remove Jetifier requirement
 * Local Payment
   * Deliver browser switch result asynchronously on main thread
 * PayPal
   * Deliver browser switch result asynchronously on main thread (fixes #500)
-* ThreeDSecure
-  * Deliver browser switch result asynchronously on main thread
 
 ## 4.9.0